pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "escrow-fx-marketplace"
version = "0.1.0"
description = "Typed FastAPI backend for an escrow FX marketplace."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.0,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.12,<1.0.0",
    "greenlet>=3.2.1,<4.0.0",
    "PyJWT>=2.10.1,<3.0.0",
    "pydantic>=2.11.1,<3.0.0",
    "pydantic-settings>=2.8.1,<3.0.0",
    "psycopg[binary]>=3.2.6,<4.0.0",
    "pwdlib[argon2]>=0.2.1,<1.0.0",
    "sqlalchemy>=2.0.39,<3.0.0",
    "uvicorn>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.21.0,<1.0.0",
    "httpx>=0.28.1,<1.0.0",
    "mypy>=1.15.0,<2.0.0",
    "pre-commit>=4.2.0,<5.0.0",
    "pytest>=8.3.5,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.11.2,<1.0.0",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "B",
    "C4",
    "E",
    "F",
    "FA",
    "I",
    "N",
    "SIM",
    "UP",
    "W",
]
ignore = [
    "E501",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
addopts = "-ra --strict-config --strict-markers"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true
pretty = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
